Move ProgressDialog and StakingDialog under components/staking/
