+++ src/components/staking/ProgressDialog.vue
