+++ src/components/staking/StakingDialog.vue
