+++ src/pages/Staking.vue
@@ -171,12 +171,12 @@ import {HMSF} from "@/utils/HMSF";
 import {waitFor} from "@/utils/TimerUtils";
 import {operatorRegistry} from "@/schemas/OperatorRegistry";
 import TransactionTableV2 from "@/components/transaction/TransactionTableV2.vue";
-import StakingDialog from "@/components/StakingDialog.vue";
+import StakingDialog from "@/components/staking/StakingDialog.vue";
 import DashboardCard from "@/components/DashboardCard.vue";
 import ConfirmDialog from "@/components/ConfirmDialog.vue";
 import {TransactionResponse} from "@hashgraph/sdk";
 import {TransactionID} from "@/utils/TransactionID";
-import ProgressDialog, {Mode} from "@/components/ProgressDialog.vue";
+import ProgressDialog, {Mode} from "@/components/staking/ProgressDialog.vue";
 import AccountLink from "@/components/values/AccountLink.vue";
 import {RewardsTransactionCache} from "@/components/staking/RewardsTransactionCache";
 import {EntityCacheStateV2} from "@/utils/EntityCacheV2";
